+++ src/widgets/WidgetSeek.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
 * VLC-Qt - Qt and libvlc connector library
-* Copyright (C) 2013 Tadej Novak <tadej@tano.si>
+* Copyright (C) 2015 Tadej Novak <tadej@tano.si>
 *
 * This library is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Lesser General Public License as published
@@ -24,10 +24,12 @@
 #if QT_VERSION >= 0x050000
     #include <QtWidgets/QHBoxLayout>
     #include <QtWidgets/QLabel>
+    #include <QtWidgets/QSlider>
     #include <QtWidgets/QProgressBar>
 #else
     #include <QtGui/QHBoxLayout>
     #include <QtGui/QLabel>
+    #include <QtGui/QSlider>
     #include <QtGui/QProgressBar>
 #endif
 
@@ -37,88 +39,87 @@
 #include "widgets/WidgetSeek.h"
 
 VlcWidgetSeek::VlcWidgetSeek(VlcMediaPlayer *player,
+                             QWidget *slider,
+                             bool connectSlider,
                              QWidget *parent)
     : QWidget(parent),
-      _vlcMediaPlayer(player)
+      _vlcMediaPlayer(player),
+      _progress(0),
+      _slider(0),
+      _connectSlider(connectSlider),
+      _labelElapsed(0),
+      _labelFull(0)
 {
-    initWidgetSeek();
+    initWidgetSeek(slider);
+}
+
+VlcWidgetSeek::VlcWidgetSeek(QWidget *slider,
+                             bool connectSlider,
+                             QWidget *parent)
+    : QWidget(parent),
+      _vlcMediaPlayer(0),
+      _progress(0),
+      _slider(0),
+      _connectSlider(connectSlider),
+      _labelElapsed(0),
+      _labelFull(0)
+{
+    initWidgetSeek(slider);
 }
 
 VlcWidgetSeek::VlcWidgetSeek(QWidget *parent)
     : QWidget(parent),
-      _vlcMediaPlayer(0)
+      _vlcMediaPlayer(0),
+      _progress(0),
+      _slider(0),
+      _connectSlider(true),
+      _labelElapsed(0),
+      _labelFull(0)
 {
-    initWidgetSeek();
+    initWidgetSeek(0);
 }
 
-VlcWidgetSeek::~VlcWidgetSeek()
-{
-    delete _seek;
-    delete _labelElapsed;
-    delete _labelFull;
-}
+VlcWidgetSeek::~VlcWidgetSeek() { }
 
-void VlcWidgetSeek::initWidgetSeek()
+void VlcWidgetSeek::initWidgetSeek(QWidget *slider)
 {
-    _lock = false;
     _autoHide = false;
 
-    _seek = new QProgressBar(this);
-    _seek->setOrientation(Qt::Horizontal);
-    _seek->setMaximum(1);
-    _seek->setTextVisible(false);
-    _seek->setMaximumHeight(15);
+    if (slider == 0)
+        slider = new QSlider();
 
-    _labelElapsed = new QLabel(this);
-    _labelElapsed->setText("00:00");
+    QAbstractSlider *sl = qobject_cast<QAbstractSlider *>(slider);
+    _slider = sl;
+    if (sl != 0 && _connectSlider) {
+        sl->setOrientation(Qt::Horizontal);
+        sl->setMaximum(1);
+        if (_vlcMediaPlayer != 0) {
+            connect(sl, SIGNAL(valueChanged(int)), _vlcMediaPlayer, SLOT(setTime(int)));
+            connect(_vlcMediaPlayer, SIGNAL(seekableChanged(bool)), sl, SLOT(setEnabled(bool)));
+        }
+    }
+    QProgressBar *bar = qobject_cast<QProgressBar *>(slider);
+    _progress = bar;
+    if (bar != 0 && _connectSlider) {
+        bar->setOrientation(Qt::Horizontal);
+        bar->setMaximum(1);
+        bar->setTextVisible(false);
+    }
 
-    _labelFull = new QLabel(this);
-    _labelFull->setText("00:00");
+    if(_labelElapsed == 0)
+        _labelElapsed = new QLabel(this);
+    _labelElapsed->setText("--:--");
 
+    if(_labelFull == 0)
+        _labelFull = new QLabel(this);
+    _labelFull->setText("--:--");
+
+    delete layout();
     QHBoxLayout *layout = new QHBoxLayout;
     layout->addWidget(_labelElapsed);
-    layout->addWidget(_seek);
+    layout->addWidget(slider);
     layout->addWidget(_labelFull);
     setLayout(layout);
-}
-
-void VlcWidgetSeek::mouseMoveEvent(QMouseEvent *event)
-{
-    event->ignore();
-
-    if (!_lock)
-        return;
-
-    updateEvent(event->pos());
-}
-
-void VlcWidgetSeek::mousePressEvent(QMouseEvent *event)
-{
-    event->ignore();
-
-    lock();
-}
-
-void VlcWidgetSeek::mouseReleaseEvent(QMouseEvent *event)
-{
-    event->ignore();
-
-    updateEvent(event->pos());
-
-    unlock();
-}
-
-void VlcWidgetSeek::wheelEvent(QWheelEvent *event)
-{
-    event->ignore();
-
-    if (!_vlcMediaPlayer)
-        return;
-
-    if (event->delta() > 0)
-        _vlcMediaPlayer->setTime(_vlcMediaPlayer->time() + _vlcMediaPlayer->length() * 0.01);
-    else
-        _vlcMediaPlayer->setTime(_vlcMediaPlayer->time() - _vlcMediaPlayer->length() * 0.01);
 }
 
 void VlcWidgetSeek::setAutoHide(bool autoHide)
@@ -135,87 +136,107 @@ void VlcWidgetSeek::setMediaPlayer(VlcMediaPlayer *player)
         disconnect(_vlcMediaPlayer, SIGNAL(timeChanged(int)), this, SLOT(updateCurrentTime(int)));
         disconnect(_vlcMediaPlayer, SIGNAL(end()), this, SLOT(end()));
         disconnect(_vlcMediaPlayer, SIGNAL(stopped()), this, SLOT(end()));
+        if (_slider != 0) {
+            disconnect(_slider, SIGNAL(valueChanged(int)), _vlcMediaPlayer, SLOT(setTime(int)));
+            disconnect(_vlcMediaPlayer, SIGNAL(seekableChanged(bool)), _slider, SLOT(setEnabled(bool)));
+        }
     }
 
     _vlcMediaPlayer = player;
+    if (player == 0)
+        return;
 
     connect(_vlcMediaPlayer, SIGNAL(lengthChanged(int)), this, SLOT(updateFullTime(int)));
     connect(_vlcMediaPlayer, SIGNAL(timeChanged(int)), this, SLOT(updateCurrentTime(int)));
     connect(_vlcMediaPlayer, SIGNAL(end()), this, SLOT(end()));
     connect(_vlcMediaPlayer, SIGNAL(stopped()), this, SLOT(end()));
+    if (_slider != 0 && _connectSlider) {
+        _slider->setOrientation(Qt::Horizontal);
+        _slider->setMaximum(1);
+        connect(_slider, SIGNAL(valueChanged(int)), _vlcMediaPlayer, SLOT(setTime(int)));
+        connect(_vlcMediaPlayer, SIGNAL(seekableChanged(bool)), _slider, SLOT(setEnabled(bool)));
+    } else if (_progress != 0 && _connectSlider) {
+        _progress->setOrientation(Qt::Horizontal);
+        _progress->setMaximum(1);
+        _progress->setTextVisible(false);
+    }
+}
+
+void VlcWidgetSeek::setSliderWidget(QWidget *slider,
+                                    bool updateSlider)
+{
+    _connectSlider = updateSlider;
+    if (slider == 0)
+        return;
+    if (slider == _slider || slider == _progress)
+        return;
+    delete _slider;
+    delete _progress;
+    initWidgetSeek(slider);
 }
 
 void VlcWidgetSeek::end()
 {
-    QTime time = QTime(0,0,0,0);
-    QString display = "mm:ss";
-
-    _labelElapsed->setText(time.toString(display));
-    _labelFull->setText(time.toString(display));
-    _seek->setMaximum(1);
-    _seek->setValue(0);
-}
-
-void VlcWidgetSeek::updateEvent(const QPoint &pos)
-{
-    if (!_vlcMediaPlayer)
-        return;
-
-    if (pos.x() < _seek->pos().x() || pos.x() > _seek->pos().x() + _seek->width())
-        return;
-
-    float click = pos.x() - _seek->pos().x();
-    float op = _seek->maximum()/_seek->width();
-    float newValue = click * op;
-
-    _vlcMediaPlayer->setTime(newValue);
-    _seek->setValue(newValue);
+    _labelElapsed->setText("--:--");
+    _labelFull->setText("--:--");
+    if (_slider != 0 && _connectSlider) {
+        _slider->setMaximum(1);
+        _slider->setValue(0);
+    } else if (_progress != 0 && _connectSlider) {
+        _progress->setMaximum(1);
+        _progress->setValue(0);
+    }
 }
 
 void VlcWidgetSeek::updateCurrentTime(int time)
 {
-    if (_lock)
-        return;
-
-    QTime currentTime = QTime(0,0,0,0).addMSecs(time);
+    QTime currentTime = QTime(0, 0);
+    currentTime = currentTime.addMSecs(time);
 
     QString display = "mm:ss";
     if (currentTime.hour() > 0)
         display = "hh:mm:ss";
 
     _labelElapsed->setText(currentTime.toString(display));
-    _seek->setValue(time);
+    if (_slider && _connectSlider) {
+        _slider->blockSignals(true);
+        _slider->setValue(time);
+        _slider->blockSignals(false);
+    } else if (_progress && _connectSlider) {
+        _progress->setValue(time);
+    }
 }
 
 
 void VlcWidgetSeek::updateFullTime(int time)
 {
-    if (_lock)
-        return;
+    if (time == 0) {
+        _labelFull->setText("--:--");
+    } else {
+        QTime fullTime = QTime(0, 0);
+        fullTime = fullTime.addMSecs(time);
 
-    QTime fullTime = QTime(0,0,0,0).addMSecs(time);
+        QString display = "mm:ss";
+        if (fullTime.hour() > 0)
+            display = "hh:mm:ss";
 
-    QString display = "mm:ss";
-    if (fullTime.hour() > 0)
-        display = "hh:mm:ss";
-
-    _labelFull->setText(fullTime.toString(display));
+        _labelFull->setText(fullTime.toString(display));
+    }
 
     if (!time) {
-        _seek->setMaximum(1);
+        if (_slider && _connectSlider)
+            _slider->setMaximum(1);
+        else if (_progress && _connectSlider)
+            _progress->setMaximum(1);
         setVisible(!_autoHide);
     } else {
-        _seek->setMaximum(time);
+        if (_slider && _connectSlider) {
+            _slider->setMaximum(time);
+            _slider->setSingleStep(1000);
+            _slider->setPageStep(time / 100);
+        }
+        if (_progress && _connectSlider)
+            _progress->setMaximum(time);
         setVisible(true);
     }
-}
-
-void VlcWidgetSeek::lock()
-{
-    _lock = true;
-}
-
-void VlcWidgetSeek::unlock()
-{
-    _lock = false;
 }